+++ app/controllers/timeline_controller.rb
@@ -1,5 +1,6 @@
 class TimelineController < ApplicationController
   def index
     @continue_draft = Memory.most_recent_incomplete
+    @years = Memory::Timeline.build
   end
 end